requirements.txt
Flask>=3.0
psycopg2>=2.9
python-dotenv>=1.0
flasgger>=0.9.7.2.dev2